api/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
influxdb-client>=1.40,<2.0
psycopg2-binary>=2.9,<3.0
pydantic>=2.0,<3.0
